giant commit part 1

+++ include/file.h
@@ -5,8 +5,8 @@
 
 typedef struct File_t
 {
-   void *stream;
-   size_t size;
+   void* stream;
+   i32 size;
 }
 File_t;
